requirements.txt
cutelog
python-dotenv
pandas
fastapi
tensorflow
keras
sqlalchemy
requests
